add easy inputing references
