+++ src/App.vue
@@ -2,18 +2,46 @@
   <div id="app">
     <img src="./assets/logo.png">
     <Flappy msg="Welcome to Your Vue.js App"/>
+    <app-header :quoteCount="quotes.length" :maxQuotes="maxQuotes"></app-header>
+    <app-new-quote @quoteAdded="newQuote"></app-new-quote>
+    <app-quote-grid :quotes="quotes" @quoteDeleted="deleteQuote"></app-quote-grid>
   </div>
 </template>
 
 <script>
-import Flappy from './components/Flappy.vue'
+  import Flappy from './components/Flappy.vue'
+  import QuoteGrid from './components/QuoteGrid.vue';
+  import NewQuote from './components/NewQuote.vue';
+  import Header from './components/Header.vue';
 
 export default {
   name: 'app',
-  components: {
-    Flappy
-  }
-}
+        data: function () {
+            return {
+                quotes: [
+                    'Just a Quote to see something'
+                ],
+                maxQuotes: 10
+            }
+        },
+        methods: {
+            newQuote(quote) {
+                if (this.quotes.length >= this.maxQuotes) {
+                    return alert('Please delete Quotes first!');
+                }
+                this.quotes.push(quote);
+            },
+            deleteQuote(index) {
+                this.quotes.splice(index, 1);
+            }
+        },
+        components: {
+            Flappy,
+            appQuoteGrid: QuoteGrid,
+            appNewQuote: NewQuote,
+            appHeader: Header
+        }
+    }
 </script>
 
 <style>
