+++ src/components/Quote.vue
