+++ src/components/QuoteGrid.vue
